+++ scripts/utils.js
@@ -13,6 +13,9 @@ export const SUPPORTED_PLACEABLES = [
   'Note',
 ];
 
+// TODO add 'Actor'
+export const UI_DOCS = ['ALL', ...SUPPORTED_PLACEABLES];
+
 export const SUPPORT_SHEET_CONFIGS = [...SUPPORTED_PLACEABLES, 'Actor', 'PlaylistSound', 'Scene'];
 
 export const SUPPORTED_HISTORY_DOCS = [...SUPPORTED_PLACEABLES, 'Scene', 'Actor', 'PlaylistSound'];